Fixed integer to float

+++ test.rb
@@ -24,7 +24,7 @@ end
 data_array = CSV.read('data.csv', headers:true)
 score_array_temp = data_array['Score']
 color_array = data_array['Color']
-
+# puts data_array
 min_color = color_array.antimode()
 max_score = score_array_temp.max
 min_score = score_array_temp.min
@@ -37,7 +37,7 @@ data_array.each do |row|
 	end
 end
 
-score_array = score_array_temp.collect{|i| i.to_i}
+score_array = score_array_temp.collect{|i| i.to_f}
 sum = score_array.inject(:+)
 puts "Average: #{(sum/ score_array.size)}"
 puts "Most Color: #{color_array.mode()}"
